requirements.txt
mido
numpy
pyFLAC
lameenc
scipy
soundfile
python-docx